Refactoring: renamed src/charset.* -> src/auth/toUtf.*
