+++ src/auth/basic/Config.cc
@@ -20,9 +20,9 @@
 #include "auth/CredentialsCache.h"
 #include "auth/Gadgets.h"
 #include "auth/State.h"
+#include "auth/toUtf.h"
 #include "base64.h"
 #include "cache_cf.h"
-#include "charset.h"
 #include "errorpage.h"
 #include "helper.h"
 #include "HttpHeaderTools.h"
